fix: formatação das datas para exibir na view

+++ app/Services/TaskService.php
@@ -2,13 +2,13 @@
 
 namespace App\Services;
 
+use Carbon\Carbon;
 use App\Repository\TaskInterface;
 
 class TaskService {
     public function __construct(protected TaskInterface $repository)
-    {
+    {}
 
-    }
     public function findAll(){
         return $this->repository->findAll();
     }
@@ -18,7 +18,9 @@ class TaskService {
     }
 
     public function new($data){
+        if($data){
         return $this->repository->new($data);
+        }
     }
 
     public function edit() {
@@ -26,10 +28,24 @@ class TaskService {
     }
 
     public function update($id, $request){
-        return $this->repository->update($id, $request);
+        if($request && $id){
+            return $this->repository->update($id, $request);
+        }
+        return 'nao atualizado';
     }
 
     public function destroy($id){
-        return $this->repository->delete($id);
+        if($id){
+            return $this->repository->delete($id);
+        }
+
+        return redirect()->back();
+    }
+
+    public function dataFormat($data){
+
+        $datas = Carbon::parse($data)->format('d/m/Y');
+
+        return $datas;
     }
 }
